requirements.txt
dagster==1.9.*
dagster-webserver==1.9.*
alembic==1.16.5
cffi==2.0.0
cryptography==46.0.7
oracledb==1.2.2
psycopg2-binary==2.9.10
pyodbc==5.1.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
typer==0.15.2
pyyaml==6.0.2
pydantic==2.9.2
